refactor(grpc-sdk,database,authentication): requested changes
