+++ libraries/grpc-sdk/src/classes/ManagedModule.ts
@@ -20,7 +20,6 @@ import convict from 'convict';
 
 export abstract class ManagedModule<T> extends ConduitServiceModule {
   private _activated: boolean = false;
-  private _migrations: boolean = false;
   protected abstract readonly configSchema?: object;
   protected abstract readonly metricsSchema?: object;
   readonly config?: convict.Config<T>;
@@ -37,14 +36,6 @@ export abstract class ManagedModule<T> extends ConduitServiceModule {
 
   get address() {
     return this._address;
-  }
-
-  get migrations() {
-    return this._migrations;
-  }
-
-  activateMigrations() {
-    this._migrations = true;
   }
 
   /**
