+++ modules/database/src/index.ts
@@ -6,7 +6,11 @@ const dbType = process.env.DB_TYPE ?? 'mongodb';
 const dbUri = process.env.DB_CONN_URI ?? 'mongodb://localhost:27017';
 
 const database = new DatabaseModule(dbType, dbUri);
-database.activateMigrations();
 const packageJsonPath = path.resolve(__dirname, '..', 'package.json');
-const moduleManager = new ModuleManager<void>(database, packageJsonPath);
+const migrationFilePath = path.resolve(__dirname, 'migrations');
+const moduleManager = new ModuleManager<void>(
+  database,
+  packageJsonPath,
+  migrationFilePath,
+);
 moduleManager.start();
